Large variety of fixes to the U-Nets
The whole idea of these changes is to remove assumption that you're only trying to predict one class (where multi-class cases automatically go for 'softmax'), which is simply not true. Although, this wasn't necessary for this project's goals, it's a solid fix to make the code more generalizable and make more sense.
# Changes
* Now, can specify out_act, the output activation function for the U-Nets
* Removed self._build_predictor because it's not that useful considering how this case is only for 2D images.
* Updated documentation